+++ CHANGELOG.md
@@ -62,6 +62,9 @@
 - [#4080](https://github.com/ChainSafe/forest/pull/4080) Fix broken
   `StateVMCirculatingSupplyInternal` RPC method on calibnet.
 
+- [#4091](https://github.com/ChainSafe/forest/pull/4091) Restore `Breeze`,
+  `Smoke`, and `Ignition` entries for calibnet
+
 - [#4093](https://github.com/ChainSafe/forest/pull/4093) Fix parsing issue in
   the `Filecoin.StateAccountKey` RPC method.